chore(meta): move TxnRequest utils to separate file
